pyproject.toml
[project]
name = "simplylab"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "Jeremy Yin", email = "jeremyyin2012@gmail.com"},
]
dependencies = [
    "fastapi>=0.110.0",
    "pymongo>=4.6.2",
    "httpx>=0.27.0",
    "uvicorn>=0.28.0",
    "openai>=1.14.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
